requirements.txt
streamlit
jinja2
requests
python-dotenv
pydantic
python-slugify
